+++ CHANGES.rst
@@ -17,6 +17,7 @@ Changelog
 * Unittest simple call of BitBar plugin.
 * Always print the raw, un-normalized version of managers, as reported by
   themselves.
+* Fetch version of all managers.
 
 
 `2.0.0 (2016-12-04) <https://github.com/kdeldycke/meta-package-manager/compare/v1.12.0...v2.0.0>`_